+++ src/command/publishWorkspace.ts
@@ -5,6 +5,7 @@ import {Logger} from '../util/logger';
 import {Constants} from './../constants';
 import {IPublishingAction} from './../spgo';
 import {UiHelper} from './../util/uiHelper';
+import { ErrorHelper } from '../util/errorHelper';
 import {SPFileService} from './../service/spFileService';
 import {AuthenticationService} from './../service/authenticationservice';
 
@@ -12,7 +13,7 @@ export default function publishWorkspace() : Thenable<any> {
     let publishingInfo : IPublishingAction = {
         fileUri : null,
         scope : Constants.PUBLISHING_MAJOR,
-        message : Constants.PUBLISHING_DEFAULT_MESSAGE
+        message : vscode.window.spgo.config.checkInMessage || Constants.PUBLISHING_DEFAULT_MESSAGE
     }
     Logger.outputMessage('Starting Workspace Publish...', vscode.window.spgo.outputChannel);
     let fileService : SPFileService = new SPFileService();
@@ -21,6 +22,6 @@ export default function publishWorkspace() : Thenable<any> {
         AuthenticationService.verifyCredentials(vscode.window.spgo, publishingInfo)
             .then((publishingInfo) => UiHelper.getPublishingMessage(publishingInfo))
             .then((publishingInfo) => fileService.uploadWorkspaceToServer(publishingInfo))
-            .catch(err => Logger.outputError(err, vscode.window.spgo.outputChannel))
+            .catch(err => ErrorHelper.handleError(err))
     );
 }