+++ src/command/resetCredentials.ts
@@ -1,7 +1,7 @@
 'use strict';
 import * as vscode from 'vscode';
 
-import {Logger} from '../util/logger';
+import { ErrorHelper } from '../util/errorHelper';
 import {AuthenticationService} from './../service/authenticationservice';
 
 //Reset the Current user's Credentials.
@@ -9,5 +9,5 @@ export default function resetCredentials() : Promise<any> {
     vscode.window.spgo.credential = null;
 
     return AuthenticationService.verifyCredentials(vscode.window.spgo)
-        .catch(err => Logger.outputError(err, vscode.window.spgo.outputChannel));
+    .catch(err => ErrorHelper.handleError(err));
 }