+++ src/command/saveFile.ts
@@ -6,6 +6,7 @@ import {Constants} from './../constants';
 import {IPublishingAction} from './../spgo';
 import {UiHelper} from './../util/uiHelper';
 import {FileHelper} from './../util/fileHelper';
+import { ErrorHelper } from '../util/errorHelper';
 import {SPFileService} from './../service/spFileService';
 import {AuthenticationService} from './../service/authenticationservice';
 
@@ -17,15 +18,15 @@ export default function saveFile(fileUri: vscode.Uri) : Thenable<any> {
         let publishAction : IPublishingAction = {
             fileUri : fileUri,
             scope : null,
-            message : Constants.PUBLISHING_DEFAULT_MESSAGE
+            message : vscode.window.spgo.config.checkInMessage || Constants.PUBLISHING_DEFAULT_MESSAGE
         }
         
         Logger.outputMessage(`Saving file:  ${fileUri.path}`, vscode.window.spgo.outputChannel);
 
         return UiHelper.showStatusBarProgress(`Saving file:  ${fileName}`,
             AuthenticationService.verifyCredentials(vscode.window.spgo, publishAction)
-                .then((publishAction) => fileService.uploadFileToServer(publishAction))  
-                .catch(err => Logger.outputError(err, vscode.window.spgo.outputChannel))
+                .then((publishAction) => fileService.uploadFileToServer(publishAction)) 
+                .catch(err => ErrorHelper.handleError(err))
         );
     }
 }